+++ src/lasers/internal.c
@@ -4,14 +4,19 @@
  * ---
  * Copyright (c) 2011-2019, Lukas Weber <laochailan@web.de>.
  * Copyright (c) 2012-2019, Andrei Alexeyev <akari@taisei-project.org>.
- */
-
-#ifndef IGUARD_laser_h
-#define IGUARD_laser_h
+*/
 
 #include "taisei.h"
 
-// TODO remove this stub header
-#include "lasers/laser.h"
+#include "internal.h"
 
-#endif // IGUARD_laser_h
+LaserInternalData lintern;
+
+void laserintern_init(void) {
+	assert(lintern.segments.num_elements == 0);
+	dynarray_ensure_capacity(&lintern.segments, 2048);
+}
+
+void laserintern_shutdown(void) {
+	dynarray_free_data(&lintern.segments);
+}
