+++ src/lasers/meson.build
@@ -1,0 +1,6 @@
+
+lasers_src = files(
+    'laser.c',
+    'draw.c',
+    'internal.c',
+)
