+++ src/util/miscmath.h
@@ -50,6 +50,8 @@ cmplxf cwmulf(cmplxf c0, cmplxf c1) attr_const;
 double cdot(cmplx c0, cmplx c1) attr_const;
 float cdotf(cmplxf c0, cmplxf c1) attr_const;
 cmplx cswap(cmplx c) attr_const;
+double ccross(cmplx a, cmplx b) attr_const;
+float ccrossf(cmplxf a, cmplxf b) attr_const;
 double psin(double) attr_const;
 double pcos(double) attr_const;
 float psinf(float) attr_const;
